Add files via upload
All of the sprites, including the drops and a quick one for the dropped powerup! The names might be a bit weird, so they can be renamed. I can also add notes on which frames of the animations they are! The slime has an idle, all three have moving frames, the slime and ghost have hurt frames, and the bat has frames for a dash attack